+++ ui/submodules/packages/components/src/components/cellRenderers/categoriesCell.tsx
@@ -1,18 +1,18 @@
 import React, { useMemo } from "react";
 import { TruncateWrapper } from "../utils";
-import { CategoryChip } from "../../common";
+import { CategoryChip, useFilterContext } from "../../common";
 import { ComponentsFiltersModel } from "../filters";
 import { CellRendererParams } from "../tableWrapper";
 
-export function CategoriesCell({ filtersContext, ...props }: CellRendererParams<ComponentsFiltersModel>): JSX.Element {
+export function CategoriesCell(props: CellRendererParams): JSX.Element {
     const { value } = props;
-    const { setFilterImmediately, getFilter } = filtersContext;
+    const { setFilter, getFilter } = useFilterContext<ComponentsFiltersModel>();
     const filterValue = useMemo(() => getFilter("CATEGORY", true), [getFilter]);
 
     return (
         <TruncateWrapper {...props}>
             {value.map((name) => (
-                <CategoryChip key={name} value={name} filterValue={filterValue} setFilter={setFilterImmediately("CATEGORY")} />
+                <CategoryChip key={name} value={name} filterValue={filterValue} setFilter={setFilter("CATEGORY")} />
             ))}
         </TruncateWrapper>
     );