+++ ui/submodules/packages/components/src/components/usages/scenarioCell.tsx
@@ -3,26 +3,32 @@ import React from "react";
 import { CellLink } from "../cellRenderers";
 import { ExternalLink, Highlight, scenarioHref } from "../../common";
 import { GridRenderCellParams } from "@mui/x-data-grid";
+import { ScenarioAvatar } from "../../scenarios/list/scenarioAvatar";
+import { ProcessType } from "nussknackerUi/components/Process/types";
+import Stack from "@mui/material/Stack";
 
 export function ScenarioCell({ filterText, ...props }: GridRenderCellParams & { filterText: string }): JSX.Element {
     const { row, value } = props;
     return (
         <CellLink component={ExternalLink} underline="hover" disabled={!value} cellProps={props} href={scenarioHref(row.id)}>
-            <Highlight value={value} filterText={filterText} />
-            <OpenInNew
-                sx={{
-                    height: ".75em",
-                    margin: ".25em",
-                    verticalAlign: "middle",
-                    opacity: 0.1,
-                    "a:hover &": {
-                        opacity: 0.5,
-                    },
-                    "a:focus &": {
-                        opacity: 0.5,
-                    },
-                }}
-            />
+            <Stack direction="row" alignItems="center">
+                <ScenarioAvatar process={row} />
+                <Highlight value={value} filterText={filterText} />
+                <OpenInNew
+                    sx={{
+                        height: ".75em",
+                        margin: ".25em",
+                        verticalAlign: "middle",
+                        opacity: 0.1,
+                        "a:hover &": {
+                            opacity: 0.5,
+                        },
+                        "a:focus &": {
+                            opacity: 0.5,
+                        },
+                    }}
+                />
+            </Stack>
         </CellLink>
     );
 }